pyproject.toml
[project]
name = "pneumonia-detection-pytorch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=12.1.1",
    "scikit-learn>=1.8.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
]
